project_code/backend/requirements.txt
# File: backend/requirements.txt
# Tech Stack: FastAPI, Aliyun Bailian

fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
dashscope==1.20.0
httpx>=0.27.0

